+++ src/Service/RequestDataConverterInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FreshAdvance\Invoice\Service;
+
+use FreshAdvance\Invoice\DataType\InvoiceConfigurationInterface;
+
+interface RequestDataConverterInterface
+{
+    public function getConfigurationFromRequest(): InvoiceConfigurationInterface;
+}